requirements.txt
flask
spacy
face_recognition
deepface
requests
beautifulsoup4
pillow
gunicorn
